Create README.md
Table 1. Multiple E3 ligases are dysregulated in XDP cells as compared to controls in general and for every cell type. 
The ones that are prevalent in both neural stem cells and induced neurons are mostly untargetable. RNF8, the only targetable gene that are prevalent in both neural stem cells and induced neurons is low in the differential expression, which raises concern for its efficacy.